[move] JwtToken 위치 이동에 따른 import문 경로 변화
